Add mock function to handle caching issue in worker environment
In this commit, a new function `mockWorkerdFetch` has been introduced. This function is used to monkey patch the `fetch` function in the worker environment because the `cache` field is not implemented there and throws an error. A note has been added for future removal once Cloudflare Workers support the `cache` field.