Move mock fetch function to web instrumentation
The mock fetch function, which is used to patch the `fetch` function in the worker environment, has been moved from 'create-reader.ts' to 'instrumentation.ts'. This was necessary because the `cache` field is not implemented in the worker environment and was causing an error. This change ensures the CMS works properly in Cloudflare until support for the