Add captcha verification to tasks
Incorporated captcha verification in task creation and update processes. A new import was added for useCaptchaToken function, and a captchaToken constant was defined. With this change, captcha token is passed as part of the data for addTaskAction and updateTaskAction functions, adding an additional layer of security for these actions. The necessary schema for captcha verification has also been defined.